# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Tasty
host: localhost:10010
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /bookmarks/{bookmark_id}:
    x-swagger-router-controller: bookmarks
    put:
      description: Update an existing bookmark
      operationId: updateBookmark
      parameters:
        - name: bookmark_id
          in: path
          description: id for the updated Bookmark
          required: true
          type: string
        - name: bookmark
          description: "Bookmark"
          required: true
          in: body
          schema:
            $ref: "#/definitions/Bookmark"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Bookmark"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /bookmarks:
    x-swagger-router-controller: bookmarks
    get:
      description: Returns a list of Bookmarks to the caller
      operationId: getBookmarks
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Bookmark"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create a new bookmark
      operationId: createBookmark
      parameters:
        - name: bookmark
          description: "Bookmark"
          required: true
          in: body
          schema:
            $ref: "#/definitions/Bookmark"
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/Bookmark"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  Bookmark:
    required:
      - title
      - url
    properties:
      notes:
        type: string
      title:
        type: string
      url:
        type: string
      tags:
        type: array
        items:
          type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string